readme and refactoring
