+++ index.html
@@ -10,7 +10,9 @@
 
   <body>
     <main id="app">
+      
     </main>
+
     <script type="module" src="/src/scripts/main.ts"></script>
   </body>
