+++ src/scripts/main.ts
@@ -1,7 +1,7 @@
 import '../styles/style.css';
 const { BASE_URL } = import.meta.env;
 
-import * as util from './utils';
+import { createCalendar } from './calendar';
 
 
 document.addEventListener( 'DOMContentLoaded', () =>
@@ -9,10 +9,11 @@ document.addEventListener( 'DOMContentLoaded', () =>
     console.log( 'Base URL:', BASE_URL );
     const main = document.querySelector( 'main' );
 
-    let calendar = util.createCalendar( 2024 );
-    console.table( calendar[ 1 ] );
+    let calendar = createCalendar( 2024 );
 
     if ( main )
     {
+
     }
+
 } );
